Object Interaction Templete
This should be the templete for the object interaction. The background image is not scaled right but the if you hover over the book area it will show the book and give you an option to click on two buttons which will prompt alerts. (the image of the book is below the background image)